handle reducer state in separate files
