+++ src/components/reducers/temp_next_program_id.js
@@ -1,0 +1,12 @@
+const initialState = 5;
+
+function tempNextProgramIDReducer(state = initialState, action) {
+  switch (action.type) {
+  case "CREATE_A_PROGRAM":
+    return state + 1;
+  default:
+    return state;
+  }
+}
+
+export default tempNextProgramIDReducer;
