+++ src/components/reducers/temp_next_workout_id.js
@@ -1,0 +1,12 @@
+const initialState = 13;
+
+function tempNextWorkoutIDReducer(state = initialState, action) {
+  switch (action.type) {
+  case "CYCLE_WORKOUT_ID":
+    return state + 1;
+  default:
+    return state;
+  }
+}
+
+export default tempNextWorkoutIDReducer;
